refactor: egg-detail-find 아이콘 라이브러리 교체 및 컴포넌트 구조 개선
- FIGMA_IMAGES 제거하고 react-native-remix-icon으로 모든 아이콘 교체
- audio-player 컴포넌트를 shared 폴더로 이동하여 재사용성 향상
- 하드코딩된 rgba 색상값을 Colors 토큰 기반으로 변경
- 사용하지 않는 스타일 제거 (audio-player 관련, badgeIcon, videoPlayIcon)
- Modal 컴포넌트 iOS 터치 이벤트 문제 수정
