+++ components/map/components/egg-detail-find/types.ts
@@ -4,20 +4,9 @@
  * Created: 2025-01-XX
  */
 
-import type { MediaType } from '@/utils/mediaType';
+import type { MediaItem } from '../shared/types';
 
 export type DiscoveryOrder = 'first' | 'second' | 'last';
-
-export interface MediaItem {
-  /** 미디어 ID */
-  id: string;
-  /** 미디어 타입 */
-  type: MediaType;
-  /** 미디어 URL */
-  url: string;
-  /** 썸네일 URL (비디오용) */
-  thumbnailUrl?: string;
-}
 
 export interface EggDiscoveryData {
   /** 이스터에그 ID */
